requirements.txt
dateutils
django-mptt>=0.2.1
django-polymorphic>=0.2
django-taggit
django-registration
django-ratings==0.3.7
django-haystack
South
django-extensions
git+git://github.com/theatlantic/django-chosen.git#egg=django-chosen
django-form-utils
